Refactor to screens folder
