+++ App.js
@@ -4,24 +4,13 @@ import React, { Component } from 'react'
 // React Native
 import { StackNavigator } from 'react-navigation'
 
-// Components
-import Layout from './components/Layout'
-import Introduction from './components/Introduction'
-import TasksLists from './components/TaskLists'
-
-class HomeScreen extends Component {
-  render () {
-    return (
-      <Layout>
-        <Introduction {...this.props} />
-      </Layout>
-    )
-  }
-}
+// Screens
+import HomeScreen from './screens/HomeScreen'
+import TasksListScreen from './screens/TaskListScreen'
 
 const HomeNavigation = StackNavigator({
   Home: { screen: HomeScreen },
-  TaskLists: { screen: TasksLists }
+  TaskLists: { screen: TasksListScreen }
 })
 
 export default class App extends Component {
